scripts/requirements.txt
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.0.1
arabic-reshaper==3.0.0
python-bidi==0.4.2
langdetect==1.0.9
requests==2.31.0 